Refactor as MM Inference Aggregator
